backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
groq==0.11.0
python-dotenv==1.0.1
